brought various Patches into single ResourcePatch object
